pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Weather App"
authors = ["Ashish"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "~0.11.3"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
sqlalchemy = "1.4.41"
psycopg2 = "^2.9.3"
fastapi = "^0.84.0"
Jinja2 = "^3.1.2"
aiofiles = "^22.1.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
